% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_mismatch}
\alias{find_mismatch}
\title{Compare a string to a vector of strings report mismatch position}
\usage{
find_mismatch(bc_to_test, all_bcs)
}
\arguments{
\item{bc_to_test}{String to check}

\item{all_bcs}{character vector to compare}
}
\description{
Compare a string to a vector of strings report mismatch position
}
