% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file}
\alias{get_file}
\title{Download a file if it doesn't exist}
\usage{
get_file(url, filename, gunzip = FALSE)
}
\arguments{
\item{url}{download url}

\item{filename}{output file name}

\item{gunzip}{uncompress file using gzip Default = FALSE, if suffix contains .gz this will
be removed.}
}
\description{
Download a file if it doesn't exist
}
