% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_hamming}
\alias{get_hamming}
\title{Compare a string to a vector of strings and report hamming distance}
\usage{
get_hamming(bc_to_test, all_bcs)
}
\arguments{
\item{bc_to_test}{String to get hamming distance}

\item{all_bcs}{character vector to compare}
}
\description{
Compare a string to a vector of strings and report hamming distance
}
