% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_hamming_pairs}
\alias{get_hamming_pairs}
\title{Compare a vector of strings to a vector of strings and report hamming distance}
\usage{
get_hamming_pairs(bcs_to_test, all_bcs)
}
\arguments{
\item{bcs_to_test}{characater vector to get hamming distance}

\item{all_bcs}{character vector to compare}
}
\description{
Compare a vector of strings to a vector of strings and report hamming distance
}
