% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_kmers}
\alias{get_kmers}
\title{get kmer-counts for character vector of sequences}
\usage{
get_kmers(seqs, n = 2L)
}
\arguments{
\item{n}{kmer size}
}
\description{
get kmer-counts for character vector of sequences
}
