% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sw}
\alias{get_sw}
\title{Compute complete striped smith waterman alignment between two dna strings (ATCG)}
\usage{
get_sw(query_seq, ref_seqs)
}
\arguments{
\item{query_seq}{query}

\item{ref_seqs}{ref}
}
\description{
Compute complete striped smith waterman alignment between two dna strings (ATCG)
}
