% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tabix.R
\name{get_tabix_chroms}
\alias{get_tabix_chroms}
\title{List chromosomes in a tabix index}
\usage{
get_tabix_chroms(filename)
}
\arguments{
\item{filename}{path to indexed tabix file}
}
\description{
List chromosomes in a tabix index
}
\examples{
tbx_file <- system.file("extdata", "tabix.bed.gz", package = "kentr")
res <- get_tabix_chroms(tbx_file)
head(res)
}
