% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getContig}
\alias{getContig}
\title{Fetch contig sequence from an indexed fasta file}
\usage{
getContig(vec, fapath)
}
\arguments{
\item{vec}{characterVector of contigs to fetch}

\item{fapath}{path to indexed fasta file}
}
\description{
Fetch contig sequence from an indexed fasta file
}
