% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getKmers}
\alias{getKmers}
\title{generate overlapping kmers from a string}
\usage{
getKmers(seqs, k = 2L)
}
\arguments{
\item{seqs}{string containing sequences to split into kmers}

\item{k}{kmer length}
}
\description{
generate overlapping kmers from a string
}
