% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSeq}
\alias{getSeq}
\title{Fetch DNA sequence from an indexed fasta file}
\usage{
getSeq(df, fapath)
}
\arguments{
\item{df}{dataframe wtih columns chrom start and end}

\item{fapath}{path to indexed fasta file}
}
\description{
Fetch DNA sequence from an indexed fasta file
}
