% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta.R
\name{gtf_to_seq}
\alias{gtf_to_seq}
\title{Extract sequences from fasta using gtf input}
\usage{
gtf_to_seq(df, fasta_path)
}
\arguments{
\item{df}{dataframe containing gtf style entries, either in as bed (chrom, start, end) or (seqnames)}

\item{fasta_path}{path to fasta file with .fai index}
}
\value{
data_frame containing transcript id column and sequence column
}
\description{
Extract sequences from fasta using gtf input
}
