% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming_distance_iterator}
\alias{hamming_distance_iterator}
\title{Compare a string to a vector of strings and report hamming distance}
\usage{
hamming_distance_iterator(bc_to_test, all_bcs)
}
\arguments{
\item{bc_to_test}{String to get hamming distance}

\item{all_bcs}{character vector to compare}
}
\description{
Compare a string to a vector of strings and report hamming distance
}
