% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_whitney.R
\name{mann_whitney}
\alias{mann_whitney}
\title{Mann whitney test}
\usage{
mann_whitney(df, id_col, alternative = "two.sided", correct = TRUE,
  mu = 0)
}
\arguments{
\item{df}{dataframe}

\item{id_col}{column in df that contains the sample ids. Test only works for two
sample comparison}

\item{alternative}{one of "two.sided", "greater", or "less" see [stats::wilcox.test]}

\item{correct}{logical see [stats::wilcox.test]}

\item{mu}{logical see [stats::wilcox.test]}
}
\description{
Mann whitney test
}
\examples{
ids <- rep(c("A", "B"), each = 1000)
nums1 <- rnorm(1000)
set.seed(42)
offset <- sample(c(-100, 100), 1)
nums2 <- nums1 + offset
df <- data.frame(ids = ids,
                 num = c(nums1, nums2))

df$num2 <- nums1 + offset + 1

mann_whitney(df, "ids")
}
