% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_inner_join}
\alias{n_inner_join}
\title{Read multiple tsv files and join by common cols}
\usage{
n_inner_join(.dir, .pattern, .cols, ...)
}
\arguments{
\item{.dir}{directory path}

\item{.pattern}{wildcard for file matching}

\item{.cols}{joins to join multiple files}

\item{...}{additional arguments to pass to [readr::read_tsv()]}
}
\description{
Read multiple tsv files and join by common cols
}
