% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read fasta format into data.frame}
\usage{
read_fasta(fasta_path)
}
\arguments{
\item{fasta_path}{filename for input fasta records}
}
\description{
Read fasta format into data.frame
}
