% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{read_genome_seq}
\alias{read_genome_seq}
\title{read a whole genome into memory as a dataframe}
\usage{
read_genome_seq(genome_file)
}
\arguments{
\item{genome_file}{path to fasta file that is indexed with samtools faidx}
}
\description{
read a whole genome into memory as a dataframe
}
\examples{
fa_path <- system.file("extdata", "test.fasta", package = "kentr")
read_genome_seq(fa_path)

}
