% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{revComp}
\alias{revComp}
\title{Reverse complement a vector of DNA sequences}
\usage{
revComp(vec)
}
\arguments{
\item{vec}{character vector of DNA [ATCGN]}
}
\description{
Reverse complement a vector of DNA sequences
}
