% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tabix.R
\name{tabix_to_df}
\alias{tabix_to_df}
\title{Read in tabix indexed file as a data.frame}
\usage{
tabix_to_df(filename = NULL, region = ".")
}
\arguments{
\item{filename}{path to tabix file}

\item{region}{samtools region query string (i.e. chr1:100-1000)

tbx_file <- system.file("extdata", "small_sorted.bam", package = "kentr")
res <- tabix_to_df(tbx_file)
head(res)

res <- tabix_to_df(tbx_file,
                   region = "chr1:3e6-3.1e6")
res}
}
\description{
Read in tabix indexed file as a data.frame
}
