% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write fasta format data to disk}
\usage{
write_fasta(df, out_path, header_col = "header", seq_col = "seq",
  gz = FALSE)
}
\arguments{
\item{df}{dataframe containing header and sequence for fasta}

\item{out_path}{filename for output fasta records}

\item{header_col}{column name containing header}

\item{seq_col}{column name containing sequences}

\item{gz}{use [R.utils::gzip] to compress output and add `.gz` suffix}
}
\description{
Write fasta format data to disk
}
