% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_gztsv}
\alias{write_gztsv}
\title{Gzip tsv output}
\usage{
write_gztsv(df, name, ...)
}
\arguments{
\item{df}{df to save with [readr::write_tsv()]}

\item{name}{output file name}

\item{...}{additional arguments passed to [readr::write_tsv()]}
}
\description{
Gzip tsv output
}
