% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sw}
\alias{get_sw}
\title{Compute complete striped smith waterman alignment between two dna strings (ATCG)}
\usage{
get_sw(query_seq, ref_seqs)
}
\arguments{
\item{query_seq}{query sequence as character vector of length one}

\item{ref_seqs}{reference sequences as a character vector}
}
\description{
Alignments will be computed for both forward and reverse complement. Forward
alignments will be indicated with strand "+", whereas reverse complement alignments will
be reported as "-". The best alignment is reported
}
